requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.12.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.40.30
botocore==1.40.30
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coverage==7.8.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
fastapi==0.115.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.43.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.113.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.52.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.10
MarkupSafe==3.0.2
narwhals==1.31.0
nodeenv==1.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.7.1
pydeck==0.9.1
pymongo==4.11.3
pyparsing==3.2.1
pyright==1.1.407
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.23.1
rsa==4.9.1
s3transfer==0.14.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.5.0
uv==0.9.6
uvicorn[standard]==0.32.0
watchdog==6.0.0
websockets==15.0.1
